compile with tsc
