+++ src/config/default.js
@@ -1,5 +1,3 @@
-"use strict";
-
 const defaultListenPort = 3000;
 
 const portFromEnv = () => {
