+++ src/config/production.js
@@ -1,5 +1,3 @@
-"use strict";
-
 module.exports = {
   electrode: {
     source: "production"
